requirements.txt
numpy==1.17.4
tensorflow-gpu==2.3.0
keras-bert==0.86.0
keras==2.4.3
tqdm==4.40.2
scikit-learn==0.19.1